Add test coverage for Redis use w/ RedisCachePool
